Fixed image loading from Stream
When Stream reports that it `CanSeek`, StreamHelper tries to load the
whole stream with a single call to Read. This may result in partial read
(e.g. stream to Azure Storage Blobs behaves like that) and corrupted
image. StreamHelper needs to read in a loop, just like with streams that
report as not seekable.